EvaluationRAG/backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
langchain-community
chromadb
langchain-chroma
ragas==0.2.10
pandas
openpyxl
python-docx
pypdf
markdown
python-dotenv
tiktoken
